+++ test/templates/landingpage/content-elements/content/text/index.js
@@ -1,6 +1,6 @@
 require('./styles.scss');
 
-const {ContentElement, PlainTextPart, Icon, cx} = require('@bsi-cx/design-build');
+const {ContentElement, PlainTextPart, Version, Icon, cx} = require('@bsi-cx/design-build');
 
 const full = require('../../../html-editor-configs/full');
 
@@ -8,6 +8,8 @@ const full = require('../../../html-editor-configs/full');
  * @type {ContentElement}
  */
 module.exports = cx.contentElement
+  .withMinVersion(Version.CX_22_0)
+  .withMaxVersion(Version.CX_22_0)
   .withElementId('title')
   .withIcon(Icon.HEADING)
   .withLabel('Titel')