+++ test/templates/landingpage/content-elements/content/title/index.js
@@ -1,11 +1,12 @@
 require('./styles.less');
 
-const {cx, Icon, ContentElement} = require('@bsi-cx/design-build');
+const {cx, Version, Icon, ContentElement} = require('@bsi-cx/design-build');
 
 /**
  * @type {ContentElement}
  */
 module.exports = cx.contentElement
+  .withMinVersion(Version.CX_1_3)
   .withElementId('content-title')
   .withIcon(Icon.HEADING)
   .withLabel('Titel')